+++ R/plantae.R
@@ -205,7 +205,7 @@
 #' @source
 #' **Associated publications:**
 #' \itemize{
-#'   \item Maestre, F.T., Benito, B.M., Berdugo, M., Concostrina-Zubiri, L., Delgado-Baquerizo, M., Eldridge, D.J., Guirado, E., Gross, N., Kefi, S., Le Bagousse-Pinguet, Y., et al. (2021). Biogeography of global drylands. \emph{New Phytologist}, 231(2), 540--558. \doi{10.1111/nph.17398}
+#'   \item Maestre, F.T., Benito, B.M., Berdugo, M., Concostrina-Zubiri, L., Delgado-Baquerizo, M., Eldridge, D.J., Guirado, E., Gross, N., Kefi, S., Le Bagousse-Pinguet, Y., et al. (2021). Biogeography of global drylands. \emph{New Phytologist}, 231(2), 540--558. \doi{10.1111/nph.17395}
 #'   \item GBIF Plantae Dataset (September 15, 2020). \doi{10.15468/dl.xh5y5g}
 #'   \item Dinerstein, E., et al. (2017). An Ecoregion-Based Approach to Protecting Half the Terrestrial Realm. \emph{BioScience}, 67(6), 534-545. \doi{10.1093/biosci/bix014}
 #'   \item Karger, D.N., et al. (2021). Climatologies at high resolution for the earth's land surface areas. EnviDat. \doi{10.16904/envidat.228.v2.1}
@@ -255,12 +255,18 @@
 #' @description
 #' Downloads and reads the extended version of the [plantae] dataset with original polygon geometries instead of point centroids, from the [spatialDataExtra](https://github.com/BlasBenito/spatialDataExtra) repository. Writes the file `plantae.gpkg` to the working directory and returns it as an `sf` dataframe.
 #' See [plantae] for details on the response variables, predictors, and data sources.
+#' Results are cached in memory for the duration of the R session; calling the function again returns the cached object instantly without re-reading from disk.
 #'
-#' 
+#'
 #' @return sf dataframe with 662 rows and 143 columns (MULTIPOLYGON geometry, WGS84).
 #' @family plantae
 #' @export
 plantae_extra <- function() {
+  cache_key <- "plantae_extra"
+  if (exists(cache_key, envir = .cache, inherits = FALSE)) {
+    return(get(cache_key, envir = .cache, inherits = FALSE))
+  }
+
   cache_dir <- tools::R_user_dir("spatialData", which = "data")
   dir.create(cache_dir, recursive = TRUE, showWarnings = FALSE)
   path <- file.path(cache_dir, "plantae.gpkg")
@@ -284,11 +290,8 @@ plantae_extra <- function() {
     )
   }
 
-  out <- sf::st_read(
-    dsn = path,
-    quiet = TRUE
-  )
-
+  out <- sf::st_read(dsn = path, quiet = TRUE)
+  assign(cache_key, out, envir = .cache)
   out
 }